Update Julia to use MMTk for perm alloc (mmtk/mmtk-julia#51)
Merge with https://github.com/mmtk/julia/pull/9.

+++ src/gc-mmtk/mmtk_julia/api/mmtk.h
@@ -34,7 +34,7 @@ extern void* alloc_large(MMTk_Mutator mutator, size_t size,
     size_t align, size_t offset, int allocator);    
 
 extern void post_alloc(MMTk_Mutator mutator, void* refer,
-    int bytes, int allocator);
+    size_t bytes, int allocator);
 
 extern void add_object_to_mmtk_roots(void *obj);
 extern void process_root_edges(closure_pointer c, void* slot);
